+++ lib/flutter_restorablez.dart
@@ -5,5 +5,7 @@ import 'dart:math' as math;
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
+part 'src/bottom_navigation_bar.dart';
+part 'src/navigation_rail.dart';
 part 'src/scroll.dart';
 part 'src/tab_bar.dart';